Improved to use Node without CommonMarkParser
