+++ Sources/NodeType.swift
@@ -6,9 +6,50 @@
 //  Copyright © 2018 Cody Winton. All rights reserved.
 //
 
+// MARK: Imports
+
 import Foundation
 
+// MARK: -
+
+struct NodeParser {
+	let type: NodeType
+	let generator: ([String]) -> Node
+}
+
+let parsers: [NodeParser] = [
+	NodeParser(type: .thematicBreak) { _ in
+		return .thematicBreak
+	},
+	NodeParser(type: .heading) { matches in
+		let level: HeadingLevel = HeadingLevel(rawValue: matches[0].count) ?? .h1
+		let raw = matches[1]
+		var components = raw.components(separatedBy: " #")
+
+		if let last = components.last?.trimmingCharacters(in: .whitespaces) {
+			let set = Set(last)
+			if set.count == 1, set.first == "#" {
+				components.removeLast()
+			}
+		}
+
+		let markdown = components.joined(separator: " #").replacingOccurrences(of: "\\#", with: "#")
+		return .heading(level: level, nodes: NodeType.heading.parse(markdown: markdown))
+	},
+	NodeParser(type: .strong) { matches in
+		return .strong(nodes: NodeType.strong.parse(markdown: matches[0]))
+	},
+	NodeParser(type: .emphasis) { matches in
+		return .emphasis(nodes: NodeType.emphasis.parse(markdown: matches[0]))
+	}
+]
+
+// MARK: -
+
 enum NodeType {
+
+	// MARK: Cases
+
 	case blockQuote
 	case code
 	case codeBlock
@@ -29,6 +70,8 @@ enum NodeType {
 	case thematicBreak
 	case customInline
 	case customBlock
+
+	// MARK: Variables
 
 	var regex: String {
 		switch self {
@@ -60,5 +103,51 @@ enum NodeType {
 			 .text, .thematicBreak, .customInline, .customBlock:
 			return []
 		}
+	}
+
+	func parse(markdown: String) -> [Node] {
+
+		var nodes: [Node] = []
+
+		guard !markdown.isEmpty else {
+			nodes.append(.text(markdown))
+			return nodes
+		}
+
+		var input = markdown
+
+		while !input.isEmpty {
+
+			var isMatched = false
+
+			for parser in parsers {
+				guard let regexMatch = input.match(regex: parser.type.regex, with: parser.type.regexTemplates) else { continue }
+				let result = parser.generator(regexMatch.captures)
+				nodes.append(result)
+
+				let inputOffset = input.index(input.startIndex, offsetBy: regexMatch.fullMatch.count)
+				input = String(input[inputOffset...])
+				isMatched = true
+				break
+			}
+
+			guard !isMatched, let first = input.first else { continue }
+
+			if first == "\n" {
+				let index = input.index(input.startIndex, offsetBy: 1)
+				nodes.append(.text(String(first)))
+				input = String(input[index...])
+			} else if self == .document, let paragraphIndex = input.index(of: "\n") {
+				let paragraph = String(input[..<paragraphIndex])
+				let paragraphNodes = NodeType.paragraph.parse(markdown: paragraph)
+				nodes.append(.paragraph(nodes: paragraphNodes))
+				input = String(input[paragraphIndex...])
+			} else {
+				nodes.append(.text(input))
+				input = ""
+			}
+		}
+
+		return nodes
 	}
 }
